chore: restructure scripts
