+++ packages/contracts/package.json
@@ -12,13 +12,13 @@
     "coverage": "yarn hardhat coverage",
 
     "// round-deploy-scripts": "//-- Round deploy Script --//",
-    "goerli-deploy-round": "yarn hardhat run scripts/deployGrantRoundImplementation.ts --network goerli",
-    "goerli-deploy-factory": "yarn hardhat run scripts/deployGrantRoundFactory.ts --network goerli",
-    "goerli-deploy-update-round": "yarn hardhat run scripts/updateGrantRoundImplementation.ts --network goerli",
-    "goerli-deploy-create-round": "yarn hardhat run scripts/createGrantRound.ts --network goerli",
+    "goerli-deploy-round": "yarn hardhat run scripts/round/deployGrantRoundImplementation.ts --network goerli",
+    "goerli-deploy-factory": "yarn hardhat run scripts/round/deployGrantRoundFactory.ts --network goerli",
+    "goerli-deploy-update-round": "yarn hardhat run scripts/round/updateGrantRoundImplementation.ts --network goerli",
+    "goerli-deploy-create-round": "yarn hardhat run scripts/round/createGrantRound.ts --network goerli",
 
     "// vote-deploy-scripts": "//-- Voting deploy Script --//",
-    "goerli-deploy-bulk-vote": "yarn hardhat run scripts/deployBulkVote.ts --network goerli"
+    "goerli-deploy-bulk-voting-mechanism": "yarn hardhat run scripts/vote/deployBulkVotingMechanism.ts --network goerli"
 
   },
   "devDependencies": {
